+++ include/dungeon.h
@@ -241,7 +241,7 @@ typedef struct mapseen {
     } flags;
     /* custom naming */
     char *custom;
-    unsigned custom_lth;
+    size_t custom_lth;
     struct mapseen_rooms {
         Bitfield(seen, 1);
         Bitfield(untended, 1);         /* flag for shop without shk */
